refactor: 增加 univer 脚本
